Avoid refreshing wxCheckBox in wxMSW if its value doesn't change
Don't call Refresh() in wxMSW when setting wxCheckBox value to the same
value it already has: this is especially important when calling Check()
or Set3StateValue() in a wxEVT_UPDATE_UI handler for it, as refreshing
it in this case results in sending another wxEVT_UPDATE_UI event, which
results in another refresh and so on, leading to very visible flicker
and high CPU usage when the application is idle.

See #24764.