nmtbcd->clrHighlightHotTrack changed as if toolbar background color is same as wxSysColourToRGB(wxSYS_COLOUR_HOTLIGHT), we would have no visual effect when hovering.
Background color is determined via m_hasBgCol ? GetBackgroundColour() : wxSystemSettings::GetColour(wxSYS_COLOUR_BTNFACE) because this also used in wxToolBar::MSWGetToolbarBgBrush() and MSWGetToolbarBgBrush is used in erase event. If we would just use GetBackgroundColour() in CDDS_ITEMPOSTPAINT phase, then the dropdown background and toolbar background are different if the toolbar background is not customized.