Make sure we generate wxEVT_QUERY_END_SESSION only once too
Because WM_QUERYENDSESSION is sent to all application TLWs, we could
generate this event multiple times, which was probably rather
unexpected, so don't do this.

Note that we rely on Windows sending us WM_ENDSESSION(FALSE) to reset
the flag set in WM_QUERYENDSESSION, but this does seem to happen in all
the current Windows versions.