Restore use of screen coordinates in wxAuiManager::ShowHint()
This function can be overridden in the application code, so it is not a
good idea to silently change its behaviour and it used to the rectangle
in screen coordinates until the changes of 29c8bfc249 (Use wxOverlay to
show docking hint instead of transparent wxFrame, 2024-01-31) when it
was changed to the rectangle in client coordinates of the window
associated with wxAuiManager.

Restore the old behaviour to avoid breaking existing code.

See #24193.