Parse HTML tag IDs to allow navigating to them later
Remember the value of ID tags attributes in the cell and find them later
when we're looking for them.

To make it easier to initialize the ID of the cells, add wxHtmlCell ctor
taking wxHtmlTag which gets the ID directly from it, and add wxHtmlTag
arguments to derived cell classes ctors too. In some cases, however, we
still need to use the new CopyId() function to set the ID of an already
existing cell.

Closes #17728.

Co-Authored-By: Vadim Zeitlin <vadim@wxwidgets.org>