Fix default background colour of ownerdrawn wxListBox items
They should use the same background as non-ownerdrawn listbox items
(i.e. wxSYS_COLOUR_LISTBOX, which white by default) but used
wxSYS_COLOUR_MENU (light grey in the default colour schema) as this is
what was used in wxOwnerDrawnBase::GetColourToUse() if no explicit
colour was specified.

Fix this by overriding GetColourToUse() in wxListBoxItem and forcing the
use of the correct colour there by default. This seems preferable to
setting the background colour to wxSYS_COLOUR_LISTBOX directly, as it
might be surprising to application code if the colour it didn't set
itself still turned out to be set.

Closes #24917.