Fix compilation of <wx/archive.h> on Clang 19 on Fedora
Clang 19.1.0 on Fedora Rawhide fails to compile <wx/archive.h>
because m_rep's members are accessed as if m_rep was not a pointer,
but it is.

This is obviously wrong with any compiler, but apparently because
wxArchiveIterator is a template class, and its operator=() is
never called in tests, its brokenness went unnoticed.

(cherry picked from commit c70a281c335d3df773b3ce4b5149aed92c9ee8a7)

See #24922.