Do offset wxClientDC and wxPaintDC by client area origin in wxMSW
We didn't use GetClientAreaOrigin() in wxMSW, unlike in all the other
ports, meaning that for the frames with a toolbar the origin of
wxPaintDC was hidden behind the toolbar itself.

Fix this and ensure that the origin is the point just below (or to the
right of) the toolbar instead.