Fix bug with unexpected events after dismissing a modal dialog
This commit fixes a long-standing problem with receiving "phantom"
events after dismissing a modal dialog shown from an event handler which
can be illustrated using the following example:

--------------------------------- >8 --------------------------------------
#include <wx/app.h>
#include <wx/button.h>
#include <wx/dialog.h>
#include <wx/frame.h>
#include <wx/msgdlg.h>
#include <wx/sizer.h>
#include <wx/stattext.h>

class TestApp : public wxApp {
public:
    bool OnInit() override {
        auto f = new wxFrame(nullptr, wxID_ANY, "wx Test");
        auto b = new wxButton(f, wxID_ANY, "Dialog");
        b->Bind(wxEVT_BUTTON, [f](wxCommandEvent&) {
            wxDialog dlg(f, wxID_ANY, "Dialog");
            dlg.SetSizerAndFit(dlg.CreateStdDialogButtonSizer(wxOK));
            dlg.ShowModal();
        });

        f->Bind(wxEVT_LEFT_UP, [f](wxMouseEvent&) {
            wxMessageBox("Got left up", "wx Test", wxOK | wxICON_INFORMATION, f);
        });

        auto* const sizer = new wxGridSizer(1);
        sizer->Add(b, wxSizerFlags().Center());
        f->SetSizer(sizer);

        f->Show();
        return true;
    }
};

wxIMPLEMENT_APP(TestApp);
--------------------------------- >8 --------------------------------------

Showing the dialog by clicking the button and then dismissing it by
clicking the button in the dialog resulted in the message box coming up.

This was due to the fact that the existing code preventing GTK from
propagating key and mouse events upwards the window parent chain based
on using the global gs_lastEvent variable didn't work if an event
handler dispatched any events from inside it, e.g. by showing a modal
dialog which ran its own nested event loop: the global variable was
overwritten by the event handling in it and, after dismissing the dialog
and returning to GTK, our handler was called again for the parent window
with the same event that we had already seen, but we didn't ignore it
any more because gs_lastEvent had been changed in the meanwhile.

Fix this by replacing the global variable with a wxEventLoop member
variable and using the currently active wxEventLoop for storing the
"last event", as it ensures that the original event is not overwritten
by event dispatching in the nested event loop.

Note that it would be tempting to achieve the same effect in a much
simpler way by connecting "after" signal handlers for all the signals
that we want to prevent from bubbling up without affecting their default
handling and this actually works fine for all of them -- except the "key
press" one, as it needs to reach the TLW parent in order to trigger the
calls to gtk_window_propagate_key_event() and gtk_window_activate_key()
there and it seems tricky to call these functions manually without
running into infinite recursion. It would be nice if we could do this,
as this would also make using our custom event source (see commit
d0406f4606 (Fix handling of identical consecutive key events,
2014-08-15) and #15802) unnecessary, but for now use this less intrusive
fix.