Fix coordinates in click events sent to window with capture
First of all, ensure that the press and release events do go to the
window with the capture, which may not have been the case at all when
the capturing window didn't have a corresponding GDK window, e.g. when
capturing mouse in wxStaticText-derived class.

And also adjust the coordinates to be correct for the captured window to
ensure that the window receives the expected events.