Fix gettext instructions in readme.txt
Write xgettext's output to .pot, not .po. Initialize a .po properly,
instead of leaving it up to the translator to set all the header fields'
values.

When updating the translation, do not use xgettext for it, as it will
overwrite all the header fields with default values. Use msgmerge instead.

Do not mention wxstd.pot at all here; keep these instructions focused on
one thing.

Do not use -C for xgettext, nor stress how important it is, as it is
remarkably unimportant here. The source file's language is correctly
auto-guessed to be C++, by the file extension.

Use msgfmt with -c, which is important to ensure valid syntax.