Define wxCharNN as typedefs for charNN_t, not wxUintNN
This makes more sense, as they're supposed to be character types, not
integer ones, and fixes the build with libc++ 19 which doesn't provide
std::char_traits specializations for types other than char, wchar_t and
charNN_t, so that std::basic_string<wxUint32> can't be compiled with it
any more without specializing std::char_traits on our own which we'd
rather avoid.

See #24959.

Closes #24958.