[wxOSX] Fix wxEvent timestamp narrowing
`wxEvent::m_timeStamp` is of type `long`, expressed in miliseconds.
On macOS this value is assigned after conversion from
`[NSEvent timestamp]` which is of type `double`, expressed in seconds.
In several such assigment instances the original value is narrowed to
`int`, apparently to silence implicit conversion warnings, see e490b0d.

However, the narrowing limits the applicability to `MAX_INT` milliseconds.
Given that int is 32 bit on mac and that `NSEvent`'s `timestamp` returns
the time in seconds since system startup, an interval between two events
can only be checked when macOS uptime is less then 24.9 days.

As macs tend to run quite stable for long periods of time without the need
for a reboot, this is a real-world problem.

Fix this by casting the source milliseconds value to `long`, which is 64
bit on mac.