Get rid of SocketPollerImpl in wxWebRequestCURL code
We used an extra level of indirection which was quite unnecessary: while
allocating platform-dependent struct responsible for polling made some
sense (although we could avoid even this), heap-allocating an object
which just contained a pointer to this struct really didn't.

So remove this object and call what was previously SocketPollerImpl just
SocketPoller directly.

No real changes, this is just a simplification and a tiny optimization.