Make wxTextCtrl::SetLabel() consistently do nothing everywhere
Previously it did the same thing as SetValue() in wxMSW but didn't do
anything in wxGTK (and probably other ports), which made it easy to
write code that only worked under Windows but mysteriously didn't show
any text under Linux.

Change it to not do anything anywhere and add an assert to help
detecting wrong uses of this function.

Also add a unit test checking this behaviour.