Use smart pointers in AUI sample serializer
This provides a better example of how to write such code as raw pointers
could have been leaked in case of exception (which never happens in the
sample but could happen in the real application code).

Also create wxXmlNode objects more consistently, using the ctor overload
not taking the parent node and adding them to the tree later.

No real changes.