Use new wxAuiPaneLayoutInfo for serialization
Using wxAuiPaneInfo for serialization was confusing as it has some
fields that must be serialized, some fields that can be serialized but
don't have to be, and some fields which should never be serialized.

Add wxAuiPaneLayoutInfo which contains just the subset of the fields
that always must be serialized to make things more clear.

Update the sample to avoid serializing fields which can never be changed
directly by the user.