Simplify deleting/removing pages from wxAuiNotebook
Use DeletePage() and RemovePage() inherited from the base class and just
override and implement DoRemovePage() instead of providing a dummy stub
for the latter.

This omits the code implementing special handling of wxAuiMDIChildFrame
but it's not really clear why did it have to be there, it looks like it
can be deleted as any other child window.