Make wxAuiTabContainer::ButtonHitTest() more convenient to use
No real changes, just return the button directly from this function
instead of using an output parameter for it.

Still preserve the old signature for compatibility, as it doesn't cost
much and it's not clear whether this function was supposed to be public
or not: lack of documentation is not a good indicator here, as none of
the functions in this class are documented, so err on the side of
caution.