Simplify handling of GREY/GRAY in wxColourDatabase
Just add the GRAY variants to the database directly instead of checking
for both GREY and GRAY when looking the colours up.

Note that we don't automatically add the version using GREY if a colour
with GRAY in it is added explicitly via AddColour() as this is not
something that had ever been documented and doesn't seem to make much
sense: application programmer using "GRAY" when adding a colour is
unlikely to use "GREY" when retrieving it later.