Fix restoring previously maximized pane when loading wxAUI layout
Setting m_hasMaximized=true, as the existing code in LoadPerspective()
did, is not enough to make the pane really maximized, and doing this
resulted in broken layout when restoring a layout with a maximized pane.

Fix this by calling MaximizePane() instead.

Note that this function only works if passed a reference to one of the
panes managed by wxAuiManager and not its copy, so we must use GetPane()
to get the proper pane to use.

Closes #24990.