Revert "Add support for multiple headers with same name to wxWebRequest"
This reverts commit 8395e5733d130a105df96f89bc66db415808b9a9 because
HTTP standard says that senders MUST NOT generate multiple lines with
the same header, with the only exception of Set-Cookie, so we shouldn't
provide a way to do it.

That commit also introduced a nasty regression with the programs calling
AddCommonHeader("User-Agent") not working at all any more after it
because WinHTTP refuses to send a request with multiple copies of this
header.

Also partially reverts 1000d67c130b08602c9fffb902ea91c3fe18355c which
was done later but is logically part of the same commit that is being
reverted.

See #24881.