Tweak wxWebView::ClearBrowsingData() API
Use invalid date instead of wxDateTime(0) to indicate that all data
should be cleared and make wxWEBVIEW_BROWSING_DATA_ALL the same as the
combination of all the other bits of this enum.

Also clear a bit more in WebKitGTK version and don't return false if
wxWEBVIEW_BROWSING_DATA_OTHER is part of the types to clear, even if it
doesn't do anything under the current platform.