Fix crash when changing wxDataViewModel items not in the tree
When sharing the same model between two or more controls it is possible
that some items are present in one tree but not the others, as the items
are only created on demand as branches are expanded, and so it is
possible to receive notifications for the items being added or deleted
even when they are not in the tree at all.

Don't crash if this happens and simply don't do anything in the notifier
functions in this case.

Note that as we now look up wxGtkTreeModelNode in the notifier itself,
this avoids multiple calls to FindNode() in GetIndexOf() and
ItemDeleted(), so this also makes deleting items a bit more efficient as
a side effect.

Closes #24148.