Fix multiple errors in the version script
Embarrassingly, almost none of the entries in the version script were
taken into account due to the function specifications not matching
anything in the generated library because they either omitted the
function signature entirely (which is apparently not allowed in extern
"C++" blocks) or specified it with the wrong "const" placement: "West
const" convention must be used here to match anything.

Fix the signatures to make things really work. Note that some of them
still don't have any effect because they apply to inline functions, but
it seems harmless to keep them.