Remove confusing "offset" from wxAuiTabContainer::RenderButtons()
This wasn't really an offset and it was reused for different things for
the left and right buttons and, finally, just doesn't seem to be
necessary at all -- using left/right buttons combined width is enough
and simpler.

Remove this variable and also move its adjustment by the tab indent size
into RenderButtons() itself, as it seems to logically belong there.

No real changes.