Use native GTK colours for wxInfoBarGeneric with GTK 3.24
This makes the generic implementation, which can be useful even in
wxGTK, as it allows adding custom controls to the info bar, more similar
to the native one in appearance and also avoids unfortunate clashes
between the tooltip colours and the native colours of other controls,
such as hyperlinks.

See #25049.

Closes #25048.