Associate standard streams with the console if we have one
Use freopen() to associate all standard streams with the console that we
attached to to let fprintf(stderr) work as well if we do have a console.

Also fall back on using stderr if WriteConsole() fails: this allows this
code to work when running under a Cygwin shell too, which wasn't the
case before.