Remove code checking that command history didn't change too
This resulted in rather unexpected behaviour when a program launched
from cmd.exe window could start by writing to this window first but then
switch to the GUI output later if another command was run from the same
window.

Just keep writing to the window which is much simpler and consistent
with how fprintf(stderr) behaves.