Fix deleting wxAuiMDIChildFrame in its Destroy()
Don't rely on wxBookCtrl::DeletePage() doing it, as it calls Destroy()
on the window since 68df81350f (Destroy, not delete, pages in
wxBookCtrlBase::DoRemovePage(), 2024-12-07), which resulted in
reentrancy and bad destruction order.

Instead, just remove the page from the book control and delete
ourselves, to ensure that we do it.

Also use FindPage() instead of iterating over the pages manually to
simplify the code.

Closes #25061, #25071.