Search parent style contexts for a background color with GTK3
'background-color' is not an inherited CSS property, but in practical terms if
a widget does not define a background color, it will have the background of the
first ancestor that does define a background. Searching up through the context
heirarchy should allow finding the correct color regardless of what form of
CSS selector a particular theme uses.
See #25048