Fix regression in wxGrid::Reparent()
Call the base class version instead of calling wxWindowBase one, which
skipped the actual reparenting that happened in the port-specific
version of this function and resulted in badly broken wxGrid appearance
when it was made child of a floating AUI pane but still appeared as a
child of main frame (because this is what it remained, natively).

This was accidentally broken in 47a4e70928 (Implement basic
accessibility support for wxGrid, 2024-03-02), see #24368.

Closes #25084.