Don't use fn_str() in calls to wx own wxMkDir() function
fn_str() should be used only when calling OS/CRT functions and using it
here resulted in needlessly constructing wxString from the C string
returned by it just to pass it to wxMkDir() which then (correctly)
called fn_str() on it again.

Skip this and just pass the string directly.