Make workaround for not creating wrong traits more compatible
The changes of 2be5f429c1 (Don't create wrong wxAppTraits during wxApp
construction, 2023-03-06) and d09245429a (Improve workaround for wrong
wxAppTraits creation, 2023-03-06) did solve the problem they were meant
to solve (see #23315 and #23319) but at the price of making an
incompatible change to GetTraits() semantics, which started returning
null pointer sometimes, whereas it always returned something valid
before, and this may have broken existing code and definitely broke
wxLog::GetActiveTarget() in wx itself which assumed the returned traits
was never null.

While we could fix it in this function itself, it seems better to always
return a valid traits object from GetTraits() and just avoid remembering
it in wxApp to ensure that the correct (overridden) CreateTraits() is
called when the application object is finally fully constructed.

The drawback of the new approach is that CreateTraits() may be called
more than once, but it seems better than risk introducing crashes into
the previously working code.

Closes #25096.