Use wxMessageOutputBest instead of Stderr in wxConsoleAppTraits
It seems better to use "best" output, which is the same as "stderr"
under Unix but can use message boxes when stderr is not available under
MSW, by default in wxConsoleAppTraits too, for consistency with what
wxLog does by default before wxApp is created.

Move wxLogOutputBest declaration from log.cpp to a private header to
allow reusing it from wxConsoleAppTraitsBase code.

This commit is best viewed with Git --color-moved option.