wxOSX: fix cursor getting reset to default on next mouse event
Fix #25131 where custom cursors get reset to default on the next mouse
event.

The first line of wxWidgetCocoaImpl::DoHandleMouseEvent is to always
call wxWidgetCocoaImpl::SetupCursor. Ultimately this goes to
wxWindow::MacSetupCursor. It looks like the final `if` was always
overwriting the choice of m_cursor with wxSTANDARD_CURSOR. Changing to
`else if` fixes the issue for me. Can't say this is the perfect fix
because there is a lot of logic happening before this point that isn't
totally clear to me.

This also shows that the cursor set logic happens on every mouse event
and thus also posts a wxSetCursor event to the app on every mouse event.
This seems suboptimal to me but is way beyond my ability to try and
change.

Closes #25134.