Make wxAuiTabContainer::TabHitTest() even more convenient to use
This replaces the more convenient TabHitTest() overloaded added recently
in 8c67124321 (Make wxAuiTabContainer::TabHitTest() more convenient to
use too, 2024-12-09) with an even more convenient one returning both the
window and its position in the tab control at once.

This allows to avoid calling GetIdxFromWindow() after a successful call
to TabHitTest() and so makes the code shorter and even more efficient,
as we don't have to iterate over the pages again in GetWindowFromIdx()
when we already have the position of the page we need in TabHitTest().

It also makes it more clear that the "index" here is actually the
physical position of the tab, i.e. corresponds to its visual order.