Don't assume tabs are dragged only in horizontal direction
This is no longer the case with wxAUI_NB_MULTILINE and the existing
logic for avoiding "snap back tab moves", which could happen because the
mouse could be on a different tab immediately after a tab move, didn't
work for dragging tabs between rows.

Instead of just checking the mouse move direction, remember the last
move position and don't allow dropping the tab into it again, at least
until it moves away from that tab.