Fix wxAuiTabContainer::Move()
This was broken by ff32a3c644 (Simplify and optimize
wxAuiTabContainer::MovePage(), 2024-12-04), we want to move the page
into the new position instead of swapping it with the page there.

Use std::rotate() to do it, which might not necessarily be more
efficient than removing and adding the page back that we did before, but
seems more clear and should hopefully avoid any off by 1 bugs.