Do not update notebook current page after tab drag
The current page index does _not_ change when its position changes and
setting the index to the position was wrong.

This partially undoes the changes of 7e530e23e1 (Update selection in
wxAuiNotebook correctly after tab move, 2016-01-11), see #16605.

Also update the sample to show the current selection when showing the
dump of all pages.