Make wxAuiNotebook::GetMainTabCtrl() public
Also return the main, and not the first, tab control from
GetActiveTabCtrl() when there is no selected page: this should be the
same thing, normally, and if it somehow isn't, returning the main tab
control makes more sense than returning some arbitrary tab control in
this case anyhow.

Simplify the code, notably removing CreateMainTabCtrl() which is not
really needed any longer.