Fix wxAuiNotebook::AdvanceSelection() when tabs are reordered
The "next" (or "previous") for the purpose of this function is the page
at the next/previous display position and not the page with the
preceding/following index, so make GetNextPage() virtual in the base
class and override it in wxAuiNotebook to do the right thing.

Also add a possibility to test AdvanceSelection() to the sample.

Closes #14454.