Don't use wxChar for iterating over wxString in wxHtmlWinParser
This doesn't work in UTF-8 build for characters outside of the BMP under
Windows, where wxChar can't hold them.

Instead, use string iterators and wxUniChar for the individual character
values.

This also simplifies the code a bit by avoiding the need for manual
memory allocation as we can let wxString do it.