Add support for wxEVT_POWER_{SUSPENDED,RESUME} under Linux
Generate these events on the systems running systemd, which provides
org.freedesktop.login1 D-Bus interface, but require acquiring system
power resource with the special new wxPOWER_DELAY "block mode" in order
to do it.

This means that existing applications need to be updated in order to
start getting these events, but it has the advantage of avoiding to have
to open the D-Bus connection on startup of every program, even if it's
never going to use it, and also forces the program to provide the
explanatory message, which is nice to have as it is shown by tools such
as systemd-inhibit to show what could be preventing system sleep.

Closes #23717.