Restore MSW look of wxAuiNotebook when not using pinned tabs
Due to the change in wxAuiNotebook code, overridden DrawTab() was not
called any longer and DrawPageTab() inherited from the base class was
used instead. Override DrawPageTab(), and also GetPageTabSize(), in
wxAuiMSWTabArt too to restore the old behaviour.

Also, because this implementation doesn't support pinned tabs (yet),
don't use it when wxAUI_NB_TAB_PIN is on.