Move wxDECLARE_CLASS_INFO_ITERATORS() from wx/object.h to rtti.h
This macro is only used in wx/rtti.h and having it defined in a
different file made it not self-contained, so move the definition to the
file which uses it.

Also remove a couple of unnecessary or redundant forward declarations.

This commit is best viewed using Git --color-moved option.