Add missing headers to wx/withimages.h
Code in this header uses wxWindow, so its forward declaration isn't
enough and the corresponding header should be included -- which
shouldn't be a problem as wxWithImages is only used in the code using
some kind of window anyhow.

Also include wx/event.h as the code here uses wxDPIChangedEvent too.