Add makefile target for checking that headers are self-contained
This can be (easily) done only when using autoconf currently, which is
not ideal, but better than never doing it at all. It also provides a
possibility to use

    $ check_headers=wx/foo.h make allheaders_selfcontained

to check the compilation of any header from the command line (note that
"check_headers" must be an environment variable and not just a make one
because versions of GNU make earlier than 4.3 have trouble with complex
expressions involving $(shell) that would be needed in this case).

Exclude a few more headers that are not supposed to be included directly
from tests/allheaders.h and add BEGIN/END STANDALONE comments to it just
to exclude the MSW-only wx/dde.h from the makefile test.