Allow serializing layout of just wxAuiNotebook
In addition to allowing to save and reload the entire wxAuiManager
layout, also allow doing it for just the given wxAuiNotebook.

This doesn't cost much, as all the required code is already present, and
just had to be refactored to extract wxAuiBookSerializer/Deserializer
classes from wxAuiSerializer/Deserializer and change wxAuiNotebook
functions to take them instead of the full [de]serializer.

Update the sample to allow editing notebook layout in XML format
interactively.