Let the system paint wxSearchButton background
For some reason, painting the bitmap created from SVG data over the
background that we draw ourselves doesn't preserve it, but not drawing
the background at all does work.

As it should be equivalent, simplify the code to do it like this, but
it's still a mystery why the background was wrong without this change.