Fix computing multi-row layout in presence of locked tabs
We can't force the page to take its "Close" button into account when
computing its size if it is locked, because the button is never shown in
this case. And, moreover, we shouldn't do it, because we could add extra
space unnecessarily if the row consists entirely of locked tabs.

So don't adjust the close button state for the locked pages and only do
it when we find the first page of another kind in LayoutMultiLineTabs().