Replace wxAUI_NB_TAB_PIN with wxAUI_NB_PIN_ON_ACTIVE_TAB
Showing "pin" icon on all tabs is too visually noisy, so (optionally)
show it only on the active tab instead.

Also change SetPageKind() to allow pinning tabs even if no special
styles are used, we can trust the program to know what it is doing and
it is common to use a tab context menu for [un]pinning the tabs instead
of using the built-in button for this, which is allowed now.