Fix install names of macOS shared libraries
Since the changes of fa2fb36a82 (Fix dylib symlinks on Apple Silicon
Macs, 2024-05-09), see #24524, the install names of the libraries were
incorrect as they were set to the name of the library file instead of
the symlink pointing to it as before. This broke backwards compatibility
as libraries linked with 3.2.5 libraries couldn't find 3.2.6 libraries
which had different names, even though they were still ABI-compatible.

Fix this by simplifying the logic of change-install-names script:
instead of iterating over all libraries, iterate only over the top-level
links, but apply the changes only to the real libraries and not the
symlinks to them.

See #25173.